refactor fetchGroup
