+++ src/containers/Home/ChatArea/ChatSection.scss
@@ -2,7 +2,7 @@
 
 .chat-wrapper {
   display: flex;
-  flex-direction: column;
+  flex-direction: column-reverse;
   justify-content: flex-start;
   font-family: $message-font;
   font-size: 1rem;
@@ -26,6 +26,8 @@
   display: flex;
   align-items: center;
   max-width: 60%;
+  margin-top: 3px;
+  margin-bottom: 3px;
 }
 
 .chat-wrapper>.user-messages {
@@ -33,11 +35,13 @@
   display: flex;
   align-items: center;
   max-width: 60%;
+  margin-top: 3px;
+  margin-bottom: 3px;
 }
 
 .chat-wrapper>.user-messages>span {
   display: block;
-  padding: 12px 16px 12px 16px;
+  padding: 10px 14px 10px 14px;
   border-radius: 48px;
   text-align: left;
   word-wrap: break-word;
